requirements.txt
aiogram==3.13.1
SQLAlchemy==2.0.36
greenlet==3.1.1
asyncpg==0.30.0
alembic==1.14.0
APScheduler==3.10.4
python-dotenv==1.0.1
tzdata==2025.2